+++ src/kerko/shortcuts.py
@@ -16,10 +16,13 @@ def data_dir() -> str:
     Get the absolute path of the data directory.
 
     If it was defined as a relative path, it will be resolved as an absolute
-    path under the app's root path.
+    path under the app's instance path, which is automatically determined by
+    Flask.
+
+    See https://flask.palletsprojects.com/en/2.3.x/config/#instance-folders.
     """
     return str(
-        pathlib.Path(current_app.root_path) / current_app.config.get(
+        pathlib.Path(current_app.instance_path) / current_app.config.get(
             'DATA_DIR',
             pathlib.Path('data') / 'kerko',
         )